Seperate Cat and Dog profiles
