+++ shared/store/src/rufferal-design/components/organism/forms/RFormEmailPassword.tsx
@@ -13,7 +13,7 @@ import {
   RFormError,
   RFormInput,
   RPasswordValidator,
-} from '../../..';
+} from '../../../..';
 
 export type AccessForm = {
   email: string;
