+++ shared/store/src/rufferal-design/components/organism/index.ts
@@ -12,9 +12,13 @@
 
 export { RAccount } from './admin/RAccount';
 export { RAllAccounts } from './admin/RAllAccounts';
+
+export { RFormCreateCat } from './forms/RFormCreateCat';
+export { RFormCreateDog } from './forms/RFormCreateDog';
+export { RFormEmailPassword } from './forms/RFormEmailPassword';
+export type { AccessForm } from './forms/RFormEmailPassword';
+
 export { RCounter } from './RCounter';
-export { RFormEmailPassword } from './RFormEmailPassword';
-export type { AccessForm } from './RFormEmailPassword';
 export { RGlobalError } from './RGlobalError';
 export { RLogoutButton } from './RLogoutButton';
 export { ROnboardingButtons } from './ROnboardingButtons';
